feat: Add validation string to session validation
The code changes include adding a new private property `validationStringKey` to the `AuthService` class. This property is used to store and retrieve a validation string from the local storage. The validation string is then included in the headers of the HTTP request made in the `validateSession` method.

Based on the recent user commits and repository commits, the commit message convention seems to be using the "chore" type for most commits. However, since this change introduces a new functionality to the code, it is more appropriate to use the "feat" type for this commit.

Please note that the commit message should not include any meta information like issue references, tags, or author names.