feat: Add CommonModule to student component imports
The code changes in the `student.component.ts` file include adding `CommonModule` to the imports of the student component. This change is necessary to use common Angular directives and pipes in the component's template. By including `CommonModule`, we ensure that the necessary functionality is available for the component to work correctly.